Added comparisons for centered interface
Very hacky - when looping over interfaces, postmodern is not included.
Instead, we render a copy of it alongside each of the other interfaces,
hard-coded, with a possibly different stylesheet each time (set in
sheets.json).

This is to allow the possibility that users of one stylesheet might be
migrated to one of several descendant sheets depending on their
interface of choice.